Temporary fix: stop webworker from leaking memory.
Drop view to 320px, it's good enough and gives decent performance

Next step is to modify the transforms to fix the origin of the plane under
the image and not the camera. This should fix distorsions when moving in the
x-y directions and height behaving in weird ways, for example not inverting
the image on negative values.

Then save user xy as offset from the center of the image, and make transform xy
as sum of offset and image center, to deliever a better UX.

After all of this add a save function, for both images and transforms.
Saving the image should save the portion of the original fullres image
contained in the view when saving

Then make it more comfortable to use, with zoom and pan handled by dragging
on the canvas, and rotation and focal length with sliders