Add points that track the image through tranformations
they will be used as vertices for 'lines' of which we can calculate the
length and output it to the view.
In this way the height doesn't need to modify the view anymore, and
it becomes easier to estimate if the scale of the transform is correct
allowing us to use absolute dimensions (in meters) in the AI module.